refactor: rename per review
