+++ datafusion/physical-plan/src/lib.rs
@@ -68,6 +68,7 @@ pub mod async_func;
 pub mod coalesce;
 pub mod coalesce_batches;
 pub mod coalesce_partitions;
+pub mod column_rewriter;
 pub mod common;
 pub mod coop;
 pub mod display;
@@ -92,7 +93,6 @@ pub mod streaming;
 pub mod tree_node;
 pub mod union;
 pub mod unnest;
-pub mod util;
 pub mod windows;
 pub mod work_table;
 pub mod udaf {
